#6029 - Enter flex mode when macromolecules mode was not entered in before and there is a drawing on canvas - refactoring
